+++ pybirdsreynolds/const.py
@@ -8,7 +8,7 @@ CANVAS_BG = None
 FILL_COLOR = None
 OUTLINE_COLOR = None
 SELECTED_INDEX = 0
-
+TRANS_HIDEN = False
 #PARAMS
 SHORT_DESCRIPTION="pybirdsreynolds is an interactive simulation of the Reynolds Boids model."
 WIDTH_CONTROLS_DEFAULT=70